Remove the "normalize" option.
The motivation is that 1. it is not part of FFTW (thus the user does
not expect it) and 2. the code to handle all cases is rather
complicated (and has never been fully written) and moreover, in many
cases, it is more efficient to use functions from Lacaml (LAPACK) but
I would prefer this library to be independent of Lacaml.